requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==3.7.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
gradio==4.44.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.8.2
numpy==1.26.4
openai==1.47.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.13.0
starlette==0.38.5
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
websockets==11.0.3
youtube-transcript-api==0.6.3
